+++ src/Products/PloneMeeting/esign/views.py
@@ -10,6 +10,7 @@ from imio.prettylink.interfaces import IPrettyLink
 from plone import api
 from Products.PloneMeeting.config import ESIGNWATCHERS_GROUP_SUFFIX
 from Products.PloneMeeting.config import MEETINGMANAGERS_GROUP_SUFFIX
+from Products.PloneMeeting.esign.utils import esign_access_groups
 from zope.i18n import translate
 
 
@@ -24,7 +25,7 @@ class PMSessionsListingView(SessionsListingView):
 
     def available(self):
         if super(PMSessionsListingView, self).available():
-            return self.tool.isManager(realManagers=True) or bool(self._get_access_groups())
+            return self.tool.isManager(realManagers=True) or bool(esign_access_groups())
 
     def get_dashboard_link(self, session):
         # if a cfg could not be initialized, we get it from the session first element
@@ -44,20 +45,14 @@ class PMSessionsListingView(SessionsListingView):
                 )
         return url
 
-    def _get_access_groups(self):
-        """Return groups of the user giving access to sessions.
-           MeetingManagers and eSign watchers have access."""
-        return self.tool.get_filtered_plone_groups_for_user(
-            suffixes=[MEETINGMANAGERS_GROUP_SUFFIX, ESIGNWATCHERS_GROUP_SUFFIX])
-
     def get_sessions(self):
         """Filter sessions by MeetingConfig.
            Only keep sessions user is MeetingManager for."""
         sessions = super(PMSessionsListingView, self).get_sessions()
         if not self.tool.isManager(realManagers=True):
-            manager_user_groups = self._get_access_groups()
+            manager_user_groups = esign_access_groups()
             manager_cfg_ids = [
-                group.replace("_%s" % MEETINGMANAGERS_GROUP_SUFFIX, "").replace(ESIGNWATCHERS_GROUP_SUFFIX, "")
+                group.replace("_%s" % MEETINGMANAGERS_GROUP_SUFFIX, "").replace("_%s" % ESIGNWATCHERS_GROUP_SUFFIX, "")
                 for group in manager_user_groups]
             sessions = [session for session in sessions if session['cfg_id'] in manager_cfg_ids]
         return sessions